backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
joblib==1.3.2
lightgbm==4.2.0
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
